% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{extract.feis}
\alias{extract.feis}
\title{Extract method for \code{feis}-class}
\usage{
extract.feis(model, include.rsquared = TRUE, include.adjrs = TRUE,
  include.nobs = TRUE, include.groups = TRUE, include.rmse = TRUE, ...)
}
\arguments{
\item{model}{an object of class \code{feis}}

\item{include.rsquared}{logical. If \code{TRUE} (default) R squared is reported.}

\item{include.adjrs}{logical. If \code{TRUE} (default) adjusted R squared is reported.}

\item{include.nobs}{logical. If \code{TRUE} number of observations is reported.}

\item{include.groups}{logical. If \code{TRUE} number of groups is reported.}

\item{include.rmse}{logical. If \code{TRUE} RMSE is reported.}

\item{...}{further arguments.}
}
\description{
Provides an extract method for usage of \code{\link[texreg]{texreg}} with \code{feis}-class.
}
\examples{
library(texreg)

setMethod("extract", signature = className("feis", "feisr"),
definition = extract.feis)

data("Produc", package = "plm")
feis1.mod <- feis("log(gsp) ~ log(pcap) | year",
                 data = Produc, id = "state", robust = TRUE)
feis2.mod <- feis("log(gsp) ~ log(pcap) + log(pc) + log(emp) + unemp | year",
                 data = Produc, id = "state", robust = TRUE)
screenreg(list(feis1.mod, feis2.mod))

}
\seealso{
\code{\link[texreg]{texreg}}, \code{\link[texreg]{screenreg}}
}
