% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feisr-package.R
\docType{package}
\name{feisr-package}
\alias{feisr-package}
\title{Estimating Fixed Effects Individual Slope Models}
\description{
The main purpose of the package `feisr` is the estimation of fixed
effects individual slopes models and respective test statistics.
The fixed effects individual slopes (FEIS) estimator is a more general
version of the well-known fixed effects estimator (FE), which allows to
control for heterogeneous slopes in addition to time-constant heterogeneity
[e.g. @Bruderl.2015.387; @Frees.2001.0; @Lemieux.1998; @Polachek.1994.0; @Ruttenauer.2020; @Wooldridge.2010.384].
This is done by running an `lm()` model on pre-transformed data, where
we (1) estimate the individual-specific predicted values for the dependent
variable and each covariate based on an individual intercept and the additional
slope variables, (2) 'detrend' the original data by these individual-specific
predicted values, and (3) run an OLS model on the residual data. The package
also provides two specification test for heterogeneous slopes.
}
\details{
The main functions of the `feisr` package are:

* `feis()`: Estimates fixed effects individual slope estimators by
applying linear `lm` models to 'detrended' data.

* `feistest()`: Estimates a regression-based Hausman test for fixed effects
individual slope models.

* `bsfeistest()`: Estimates a bootstrapped Hausman test for fixed effects
individual slope models.

The functions included in the R package `feisr` are also available in the
[xtfeis ado](https://www.stata.com/meeting/germany10/germany10_ludwig.pdf)
for Stata. The package [plm](https://CRAN.R-project.org/package=plm) provides
esimation of related models, as for instance the mean group (MG) or common
correlated effects mean groups (CCEMG) estimator via `pmg()` function and
the estimator of models with variable coefficients via `pvcm()`.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[plm]{plm}}, \code{\link[plm]{pvcm}}, \code{\link[plm]{pmg}}
}
\author{
Tobias Ruettenauer

Volker Ludwig
}
