% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feistest.R
\name{feistest}
\alias{feistest}
\title{Augmented Regression Test}
\usage{
feistest(model = NA, robust = FALSE, type = c("both", "art1", "art2"),
  ...)
}
\arguments{
\item{model}{an object of class "\code{feis}".}

\item{robust}{logical. If \code{TRUE} uses cluster robust standard errors (Default is \code{FALSE}).}

\item{type}{one of "\code{both}" (the Default), "\code{art1}", "\code{art2}".}

\item{...}{further arguments.}
}
\value{
An object of class "\code{feistest}", containing the following elements:
\item{wald_feis}{an object of class "\code{wald.test}" testing the fixed effects individual slopes model
  against the conventional fixed effects model.}
\item{wald_fe}{an object of class "\code{wald.test}" testing the fixed effects model
  against the random effects model.}
\item{vcov1}{the variance-covariance matrix of CREIS.}
\item{vcov2}{the variance-covariance matrix of CRE.}
\item{CREIS}{an object of class "\code{plm}" estimating a Correlated Random Effect Individual Slope model.}
\item{CRE}{an object of class "\code{plm}" estimating a Correlated Random Effect model.}
\item{call}{the matched call.}
\item{robust}{logical. If \code{TRUE} cluster robust standard errors were used.}
\item{formula}{an object of class "\code{Formula}" describing the model.}
\item{type}{the type of performed test(s).}
}
\description{
Estimates a regression-based Hausmann test for fixed effects individual slope models.
}
\details{
The Hausmann test can be computed by estimating a correlated random effects model
\insertCite{@see @Wooldridge.2010.384, pp. 328-334}{feisr}. This is achieved by
estimating a Mundlak \insertCite{Mundlak.1978.0}{feisr} specification using random effects models
with \code{\link[plm]{plm}}.
Subsequently, \code{feistest} tests whether the time-constant variables / slope variables are correlated with
the unobserved heterogeneity by using a Wald-test with \code{\link[aod]{wald.test}}.

While \code{type="art2"} estimates the conventional regression-based Hausmann test
\insertCite{@as described in @Wooldridge.2010.384, pp. 328-334}{feisr} comparing conventional
fixed effects models against random effects models, \code{type="art1"} estimates an extended
regression-based Hausmann test comparing fixed effects individual slope models and conventional
fixed effects model. For "\code{art1} the Mundlak-specification includes the person-specific averages,
but additionally the person-specific slope estimates used for "detrending" in \code{\link[feisr]{feis}}.
The Wald test of \code{type="art1"} is applied to the slope variables only.

If specified, \code{feistest} uses panel-robust standard errors.
}
\examples{
data("Produc", package = "plm")
feis.mod <- feis(log(gsp) ~ log(pcap) + log(pc) + log(emp) + unemp | year,
                 data = Produc, id = "state", robust = TRUE)
ht <- feistest(feis.mod, robust = TRUE, type = "both")
summary(ht)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[feisr]{feis}}, \code{\link[plm]{plm}}, \code{\link[aod]{wald.test}}, \code{\link[plm]{phtest}}
}
